Reshuffle of rotation/gridding utils
