+++ diffsims/utils/gridding_utils.py
@@ -24,76 +24,9 @@ import numpy as np
 from itertools import product
 from transforms3d.euler import axangle2euler, euler2axangle, euler2mat
 from transforms3d.quaternions import quat2axangle, axangle2quat, mat2quat, qmult
-from diffsims.utils.rotation_conversion_utils import vectorised_euler2axangle, vectorised_quat2axangle, vectorised_axangle2euler
+from diffsims.utils.rotation_conversion_utils import *
 
 
-def convert_axangle_to_correct_range(vector, angle):
-    """
-    This repo uses axis-angle pairs between (0,pi) - however often wider
-    ranges are used, most common are (0,2pi) and (-pi,pi), this function corrects
-    for these
-
-    Parameters
-    ----------
-    vector : iterable of length 3
-        [x,y,z] of the axis
-    angle : float
-        in radians
-
-    Returns
-    -------
-    vector, angle :
-        correct forms of the inputs
-
-    See Also
-    --------
-    vectorised_axangle_to_correct_range : for fast processing of bigger inputs
-
-    """
-    if (angle >= 0) and (angle < np.pi):  # input in the desired convention
-        pass
-    elif (angle >= -np.pi) and (angle < 0):
-        vector = np.multiply(vector, -1)
-        angle = angle * -1
-    elif (angle >= np.pi) and (angle <= 2 * np.pi):
-        vector = np.multiply(vector, -1)
-        angle = 2 * np.pi - angle
-    else:
-        raise ValueError("You have an axis-angle angle outside of acceptable ranges:`{}`".format(angle))
-
-    return vector, angle
-
-def vectorised_axangle_to_correct_range(data):
-    """
-    This repo uses axis-angle pairs between (0,pi) - however often wider
-    ranges are used, most common are (0,2pi) and (-pi,pi), this function corrects
-    for these
-
-    Parameters
-    ----------
-    data : (N,4)
-        axangles
-
-    Returns
-    -------
-    data : (N,4)
-        Corrected forms of the input
-
-    See Also
-    --------
-    convert_axangle_to_correct_range : for single data items
-    """
-
-    # second clause in unvectorised
-    data[:,3] = np.where(np.logical_and(data[:,3] >= -np.pi,data[:,3] < 0),-data[:,3],data[:,3])
-
-    # third clause in unvectorised
-    third_case_truth = np.logical_and(data[:,3] >= np.pi, data[:,3] <= 2 * np.pi)
-    for i in [0,1,2]: # third clause part 1
-        data[:,i] = np.where(third_case_truth,-data[:,i],data[:,i])
-    data[:,3] = np.where(third_case_truth,2*np.pi - data[:,3],data[:,3]) # third clause part 2
-
-    return data
 
 def vectorised_qmult(q1, qdata):
     """ A vectorised implementation that multiplies qdata (array) by q1 (single quaternion) """
@@ -126,152 +59,7 @@ def _get_rotation_to_beam_direction(beam_direction):
     alpha,beta,gamma = axangle2euler(axis,angle,'szxz')
     return np.rad2deg(alpha),np.rad2deg(beta)
 
-class AxAngle():
-    """
-    Class storing rotations in the axis-angle convention. Each row reads
-    as [vx,vy,vz,theta], where [vx,vy,vz] is the rotation axis (normalised)
-    and theta is the rotation angle in radians in range (0,pi]
-    """
 
-    def __init__(self, data):
-        self.data = data.astype('float')
-        # apply identity rotation change
-        self._check_data()
-        return None
-
-    def _check_data(self):
-        """ Checks the data within AxAngle is acceptable, having the correct dimensions,
-        acceptable angles and normalised vectors """
-        if self.data.shape[1] != 4:
-            raise ValueError("Your data is not in the correct shape")
-        if np.any(self.data[:, 3] < 0) or np.any(self.data[:, 3] > np.pi):
-            raise ValueError("Some of your angles lie outside of the range (0,pi)")
-        if not np.allclose(np.linalg.norm(self.data[:, :3][self.data[:, 3] > 0], axis=1), 1):
-            raise ValueError("You no longer have normalised direction vectors")
-        # drop the angle zero exclusion on the above for identity rotations
-        return None
-
-    def remove_large_rotations(self, threshold_angle):
-        """
-        Removes rotations that above a threshold angle
-
-        Parameters
-        ----------
-        thereshold_angle : float
-            angle in radians, rotations larger than this are removed
-
-        Returns
-        -------
-        None :
-            This functions operates in place
-        """
-        self._check_data()
-        self.data = self.data[self.data[:, 3] < threshold_angle]
-        return None
-
-    def to_Euler(self, axis_convention):
-        """
-        Produces euler angles from the axis-angle pairs.
-
-        Parameters
-        ----------
-        axis_convention : str
-            transforms3d compliant euler string
-
-        Returns
-        -------
-        out_eulers : diffsims.Euler
-        """
-        self._check_data()
-        eulers = vectorised_axangle2euler(self.data,axis_convention)
-        eulers = np.rad2deg(eulers)
-        return Euler(eulers, axis_convention)
-
-    def to_Quat(self):
-        """ A lightweight port of transforms3d functionality, vectorised"""
-        self._check_data() #means that our vectors need not be checked for normalisation
-        vector = self.data[:,:3]
-        t2 = self.data[:,3] / 2.0
-        st2 = np.sin(t2)
-        w = np.cos(t2).reshape(self.data.shape[0],1)
-        xyz = np.multiply(vector,st2.reshape(self.data.shape[0],1))
-        return np.hstack((w,xyz))
-
-
-    @classmethod
-    def from_Quat(cls, data):
-        # TODO: input checking
-        axangles = vectorised_quat2axangle(data)
-        # needs a vectorised version of convert_axangle_to_correct_range
-        return AxAngle(axangles)
-
-
-class Euler():
-    """
-    Class storing rotations as euler angles.
-    Each row reads as [alpha,beta,gamma], where alpha, beta and gamma are rotations
-    in degrees around the axes specified by Euler.axis_conventions
-    as defined in transforms3d. Please always remember that Euler angles are difficult.
-    """
-
-    def __init__(self, data, axis_convention='rzxz'):
-        self.data = data.astype('float')
-        self.axis_convention = axis_convention
-        self._check_data()
-        return None
-
-    def _check_data(self):
-        """ Checks data within Euler is acceptable, to be used at the start
-        of methods """
-        if self.data.shape[1] != 3:
-            raise ValueError("Your data is not in the correct shape")
-        if np.any(self.data[:] > 360):
-            raise ValueError("Some of your angles are greater 360")
-
-        return None
-
-    def to_AxAngle(self):
-        """ Converts an Euler object to an AxAngle object
-
-        Returns
-        -------
-        axangle : diffsims.AxAngle object
-        """
-        self._check_data()
-        self.data = np.deg2rad(self.data)  # for the transform operation
-
-        if self.axis_convention == 'rzxz' or self.axis_convention =='szxz':
-            stored_axangle = vectorised_euler2axangle(self.data, axes=self.axis_convention)
-            stored_axangle = vectorised_axangle_to_correct_range(stored_axangle)
-
-        else:
-            # This is very slow
-            stored_axangle = np.ones((self.data.shape[0], 4))
-            for i, row in enumerate(self.data):
-                temp_vect, temp_angle = euler2axangle(row[0], row[1], row[2], self.axis_convention)
-                temp_vect, temp_angle = convert_axangle_to_correct_range(temp_vect, temp_angle)
-                for j in [0, 1, 2]:
-                    stored_axangle[i, j] = temp_vect[j]
-                    stored_axangle[i, 3] = temp_angle  # in radians!
-
-        self.data = np.rad2deg(self.data)  # leaves our eulers safe and sound
-        return AxAngle(stored_axangle)
-
-    def to_rotation_list(self,round_to=2):
-        """
-        Parameters:
-
-        round_to : int or None
-
-        """
-        if round_to is not None:
-            round = np.round(self.data,2)
-        else:
-            round = self.data
-
-        starter_list = round.tolist()
-        tuple_list  = [tuple(x) for x in starter_list]
-        return tuple_list
 
 def rotation_matrix_from_euler_angles(euler_angles):
     """
@@ -367,135 +155,3 @@ def _create_advanced_linearly_spaced_array_in_rzxz(resolution, max_alpha, max_be
     gamma = np.linspace(0, max_gamma, num=steps_gamma, endpoint=False)
     z = np.asarray(list(product(alpha, beta, gamma)))
     return Euler(z, axis_convention='rzxz')
-
-
-""" Fundemental Zone Functionality """
-
-
-def get_proper_point_group_string(space_group_number):
-    """
-    Parameters
-    ----------
-    space_group_number : int
-
-    Returns
-    -------
-    point_group_str : str
-        The proper point group string in --- convention
-
-    Notes
-    -----
-    This function enumerates the list on https://en.wikipedia.org/wiki/List_of_space_groups
-    Point groups (32) are converted to proper point groups (11) using the Schoenflies
-    representations given in that table.
-    """
-    if space_group_number in [1, 2]:
-        return '1'  # triclinic
-    if 2 < space_group_number < 16:
-        return '2'  # monoclinic
-    if 15 < space_group_number < 75:
-        return '222'  # orthorhomic
-    if 74 < space_group_number < 143:  # tetragonal
-        if (74 < space_group_number < 89) or (99 < space_group_number < 110):
-            return '4'  # cyclic
-        else:
-            return '422'  # dihedral
-    if 142 < space_group_number < 168:  # trigonal
-        if 142 < space_group_number < 148 or 156 < space_group_number < 161:
-            return '3'  # cyclic
-        else:
-            return '32'  # dihedral
-    if 167 < space_group_number < 194:  # hexagonal
-        if 167 < space_group_number < 176 or space_group_number in [183, 184, 185, 186]:
-            return '6'  # cyclic
-        else:
-            return '622'  # dihedral
-    if 193 < space_group_number < 231:  # cubic
-        if 193 < space_group_number < 207 or space_group_number in [215, 216, 217, 218, 219, 220]:
-            return '432'  # oct
-        else:
-            return '23'  # tet
-
-
-def axangle2rodrigo_frank(z):
-    # converts to [vx,vy,vz,RF]
-    # RF = tan(omega/2)
-    z[:, 3] = np.tan(np.divide(z[:, 3], 2))
-    return z
-
-
-def rodrigo_frank_to_axangle():
-    pass
-
-
-def numpy_bounding_plane(data, vector, distance):
-    """
-
-    Raises
-    -----
-    ValueError : This function is unsafe if pi rotations are preset
-    """
-    if not np.all(np.is_finite(data)):
-        raise ValueError("pi rotations, be aware")
-
-    return data
-
-
-def cyclic_group(data, order):
-    """ By CONVENTION the rotation axis is the cartesian z axis
-    Note: Special case, as pi rotations are present we avoid a call to numpy_bounding_plane"""
-    z_distance = np.multiply(data[2], data[3])
-    z_distance = np.abs(np.nan_to_num(z_distance))  # case pi rotation, 0 z component of vector
-    return data[z_distance < np.tan(np.pi / order)]
-
-
-def dihedral_group(data, order):
-    pass
-
-
-def octahedral_group(data):
-    pass
-
-
-def tetragonal_group(data):
-    for direction in [(1, 1, 1), (1, 1, -1)]:  # etc
-        data = numpy_bounding_plane()
-
-
-def rf_fundemental_zone(axangledata, point_group_str):
-    rf = axangle2rodrigo_frank(axangledata)
-    if point_group_str in ['1', '2', '3', '4', '6']:
-        rf = cyclic_group(rf, order=int(point_group_str))
-    elif point_group_str in ['222', '32', '422', '622']:
-        rf = dihedral_group(rf, order=int(point_group_str[0]))
-    elif point_group_str == '23':
-        rf = tetragonal_group(rf)
-    elif point_group_str == '432':
-        rf = octahedral_group(rf)
-    return rodrigo_frank_to_axangle(rf)
-
-
-def reduce_to_fundemental_zone(data, fundemental_zone):
-    """
-    Parameters
-    ----------
-    data : diffsims.AxAngle
-
-    fundemental_zone : str
-        A proper point group, allowed values are:
-            '1','2','222','4','422','3','32','6','622','432','23'
-
-    Returns
-    -------
-    reduced_data : diffsims.AxAngle
-
-    """
-
-    """
-    Use a conditional .remove_large_angles() here for speed.
-    we know what are max angles are, so save some time by cutting out chunks
-    see Figure 5 of "On 3 dimensional misorientation spaces"
-    """
-    # convert to rodrigo-frank
-    # call FZ functionality
-    return None
